% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_state_water_data.R
\name{load_state_water_data}
\alias{load_state_water_data}
\title{Load State Water Data}
\usage{
load_state_water_data()
}
\value{
A data frame containing the state water data.
}
\description{
This function loads the \code{state_water_data_first_1000.xlsx} file included in the \code{statewaterdata} package.
}
\examples{
data <- load_state_water_data()
}
